// src/views/payroll/employee/employee.type.ts
export interface EmployeeType {
  id: number;
  nsifNumber: string;
  hireDate: any;
  terminationDate: any;
  employmentStatus: string;
  employmentType: string;
  baseSalary: number | string | null;
  baseSalaryF: number | string | null;
  department: any;
  departmentId: any;
  position: any;
  positionId: any;
  personnel: any;
  personnelId: any;
  payType: string;
  hourlySalary: number | string | null;
  hourlySalaryF: number | string | null;
}